Solved day 5.

+++ config.h
@@ -12,6 +12,6 @@
 #define CONFIG_H
 
 #define SIGNATURE "Written by Abed Na'aran, compiled on " __DATE__ ".\n"
-typedef int ResultType;
+typedef long ResultType;
 
 #endif /* CONFIG_H */
